requirements.txt
fastapi==0.111.0
onnxruntime-gpu==1.18.0
piper-tts==1.2.0
pydantic_settings==2.3.4
pyyaml==6.0.1
requests==2.31.0
uvicorn==0.29.0
